locale.c: Rewrite localeconv() handling
localeconv() returns a structure contaiing fields that are associated
with two different categories: LC_NUMERIC and LC_MONETARY.  Perl via
POSIX::localeconv() reutrns a hash containing all the fields.

Testing on Windows showed that if LC_CTYPE is not the same locale as
LC_MONETARY for the monetary fields, or isn't the same as LC_NUMERIC for
the numeric ones, mojibake can result.

The solution to similar situations elsewhere in the code is to toggle
LC_CTYPE into being the same locale as the one for the returned fields.
But those situations only have a single locale that LC_CTYPE has to
match, so it doesn't work here when LC_NUMERIC and LC_MONETARY are
different locales.   Unlike Schrödinger's cat, LC_CTYPE has to be one or
the other, not both at the same time.

The previous implementation did not consider this possibility, and
wasn't easily changeable to work.

Therefore, this rewrites a bunch of it.  The solution used is to call
localeconv() twice, once for the LC_NUMERIC locale (with LC_CTYPE
corresponding), and once for LC_MONETARY.  If those locales are the
same, the second call is collapsed into the first one, resulting in just
one call.  This likely scenario complicated the code so as to avoid
unnecessary calls.

Another complication is that on platforms that lack nl_langinfo(),
(Windows, for example), localeconv() is used to emulate portions of it.
Previously there was a separate function to handle this, using an SV()
cast as an HV() to avoid using a hash that wasn't actually necessary.
That proved to lead to extra duplicated code under the new scheme, so
that function was collapsed into a single one and a real hash is used in
all circumstances, but is only populated with the one or two fields
needed for the emulation.

The only part of this commit that I thought could be split off from the
rest concerns the fact that localeconv()'s return is not thread-safe,
and so must be copied to a safe place (the hash) while in a critical
section, locking out all other threads.  Before this commit, that
copying was accompanied by determining if each string field needed to be
marked as UTF-8.  That determination isn't necessarily trivial, and in
fact can be deferred until out of the critical section.  This commit
does that.  And, with some effort, that part could have been split into
a separate commit. but I didn't think it was worth the effort.