Fixup UTF-8 overflow detection on 32bit words
On a 32-bit machine, any UTF-8 sequence that begins with either \xFE or
\xFF is for a code point that won't fit in the available bits, unless
that sequence is also actually an overlong representation of a smaller
code point.  Prior to this commit, the possibility of it being an
overlong was given undue weight, so overflow messages were suppressed.

This is pretty obscure.  This happens only when other malformations are
present, and came to light due to expanded testing that is coming in the
next commit.  But our goal has been to diagnose all the malformations
present for any given sequence; not just a single malformation.