Actually deprecate my_sprintf()
This has been considered deprecated since before 5.32.  But since it is
implemented as a macro, nothing notified anyone using it.  This commit
turns it into a function, and the compiler should warn of any calls to
it.

There is #ifdef'd out code in utf8.c that could be adapted to output a
run-time warning message for the first call from every call-site to this
function, giving more warning.